fixed unsigned integer issue

+++ Pod/Classes/JAQDiceView.m
@@ -146,7 +146,7 @@
 	}
 }
 
-- (NSUInteger)boxUpIndex:(SCNNode *)boxNode {
+- (int)boxUpIndex:(SCNNode *)boxNode {
 	GLKVector3 rotatedUp = [self rotatedVector:boxNode];
 	GLKVector3 boxNormals[6] = {
 		{{0,-1,0}},
